# do a bunch of optimizations (like switching to Neri-Schneider)