--- a/src/fmt/util.rs
+++ b/src/fmt/util.rs
@@ -141,6 +141,7 @@ impl Decimal {
 
     /// Returns the total number of ASCII bytes (including the sign) that are
     /// used to represent this decimal number.
+    #[inline]
     const fn len(&self) -> u8 {
         self.end - self.start
     }
@@ -148,11 +149,13 @@ impl Decimal {
     /// Returns the ASCII representation of this decimal as a byte slice.
     ///
     /// The slice returned is guaranteed to be valid ASCII.
+    #[inline]
     pub(crate) fn as_bytes(&self) -> &[u8] {
         &self.buf[usize::from(self.start)..usize::from(self.end)]
     }
 
     /// Returns the ASCII representation of this decimal as a string slice.
+    #[inline]
     pub(crate) fn as_str(&self) -> &str {
         // SAFETY: This is safe because all bytes written to `self.buf` are
         // guaranteed to be ASCII (including in its initial state), and thus,
@@ -311,9 +314,9 @@ impl Fractional {
 /// misnomer, but the range of possible values is still correct. (That is, the
 /// fractional component of an hour is still limited to 9 decimal places per
 /// the Temporal spec.)
-#[inline(never)]
+#[inline(always)]
 pub(crate) fn parse_temporal_fraction<'i>(
-    mut input: &'i [u8],
+    input: &'i [u8],
 ) -> Result<Parsed<'i, Option<t::SubsecNanosecond>>, Error> {
     // TimeFraction :::
     //   TemporalDecimalFraction
@@ -344,41 +347,48 @@ pub(crate) fn parse_temporal_fraction<'i>(
     // DecimalDigit :: one of
     //   0 1 2 3 4 5 6 7 8 9
 
+    #[inline(never)]
+    fn imp<'i>(
+        mut input: &'i [u8],
+    ) -> Result<Parsed<'i, Option<t::SubsecNanosecond>>, Error> {
+        let mkdigits = parse::slicer(input);
+        while mkdigits(input).len() <= 8
+            && input.first().map_or(false, u8::is_ascii_digit)
+        {
+            input = &input[1..];
+        }
+        let digits = mkdigits(input);
+        if digits.is_empty() {
+            return Err(err!(
+                "found decimal after seconds component, \
+                 but did not find any decimal digits after decimal",
+            ));
+        }
+        // I believe this error can never happen, since we know we have no more
+        // than 9 ASCII digits. Any sequence of 9 ASCII digits can be parsed
+        // into an `i64`.
+        let nanoseconds = parse::fraction(digits, 9).map_err(|err| {
+            err!(
+                "failed to parse {digits:?} as fractional component \
+                 (up to 9 digits, nanosecond precision): {err}",
+                digits = escape::Bytes(digits),
+            )
+        })?;
+        // I believe this is also impossible to fail, since the maximal
+        // fractional nanosecond is 999_999_999, and which also corresponds
+        // to the maximal expressible number with 9 ASCII digits. So every
+        // possible expressible value here is in range.
+        let nanoseconds =
+            t::SubsecNanosecond::try_new("nanoseconds", nanoseconds).map_err(
+                |err| err!("fractional nanoseconds are not valid: {err}"),
+            )?;
+        Ok(Parsed { value: Some(nanoseconds), input })
+    }
+
     if input.is_empty() || (input[0] != b'.' && input[0] != b',') {
         return Ok(Parsed { value: None, input });
     }
-    input = &input[1..];
-
-    let mkdigits = parse::slicer(input);
-    while mkdigits(input).len() <= 8
-        && input.first().map_or(false, u8::is_ascii_digit)
-    {
-        input = &input[1..];
-    }
-    let digits = mkdigits(input);
-    if digits.is_empty() {
-        return Err(err!(
-            "found decimal after seconds component, \
-             but did not find any decimal digits after decimal",
-        ));
-    }
-    // I believe this error can never happen, since we know we have no more
-    // than 9 ASCII digits. Any sequence of 9 ASCII digits can be parsed
-    // into an `i64`.
-    let nanoseconds = parse::fraction(digits, 9).map_err(|err| {
-        err!(
-            "failed to parse {digits:?} as fractional component \
-             (up to 9 digits, nanosecond precision): {err}",
-            digits = escape::Bytes(digits),
-        )
-    })?;
-    // I believe this is also impossible to fail, since the maximal
-    // fractional nanosecond is 999_999_999, and which also corresponds
-    // to the maximal expressible number with 9 ASCII digits. So every
-    // possible expressible value here is in range.
-    let nanoseconds = t::SubsecNanosecond::try_new("nanoseconds", nanoseconds)
-        .map_err(|err| err!("fractional nanoseconds are not valid: {err}"))?;
-    Ok(Parsed { value: Some(nanoseconds), input })
+    imp(&input[1..])
 }
 
 /// This routine returns a span based on the given with fractional time applied