--- a/src/lib.rs
+++ b/src/lib.rs
@@ -778,5 +778,6 @@ mod tests {
         dbg!((t::SpanMilliseconds::MIN, t::SpanMilliseconds::MAX));
         dbg!((t::SpanMicroseconds::MIN, t::SpanMicroseconds::MAX));
         dbg!((t::SpanNanoseconds::MIN, t::SpanNanoseconds::MAX));
+        dbg!((t::UnixSeconds::MIN, t::UnixSeconds::MAX));
     }
 }