This PR completely re-organizes Jiff's benchmarks and implements a
bunch of optimizations.

I've said on a number of occasions that I hadn't yet paid too much
attention to optimizing various code paths in Jiff. I had done some
work in various areas, but nothing like what I did here over the past
week. There are some major speed-ups for some operations:

```
$ critcmp base update -g '(.*)/(?:jiff)$'
group                                                           base//jiff                             update//jiff
-----                                                           ----------                             ------------
civil_datetime/add_days/diffyear/duration                       2.12     31.0±0.25ns        ? ?/sec    1.00     14.6±0.29ns        ? ?/sec
civil_datetime/add_days/diffyear/span                           4.08     88.8±0.36ns        ? ?/sec    1.00     21.8±0.20ns        ? ?/sec
civil_datetime/add_days/sameyear/duration                       2.13     31.0±0.28ns        ? ?/sec    1.00     14.6±0.23ns        ? ?/sec
civil_datetime/add_days/sameyear/span                           4.08     88.9±0.39ns        ? ?/sec    1.00     21.8±0.15ns        ? ?/sec
civil_datetime/to_datetime_static/zoneinfo                      1.20     19.5±0.22ns        ? ?/sec    1.00     16.2±0.17ns        ? ?/sec
civil_datetime/to_timestamp_tzdb_lookup/zoneinfo                3.00    122.0±0.93ns        ? ?/sec    1.00     40.7±0.17ns        ? ?/sec
date/add_days/diffyear/duration                                 3.18     20.8±0.16ns        ? ?/sec    1.00      6.5±0.04ns        ? ?/sec
date/add_days/diffyear/span                                     4.86     31.7±0.30ns        ? ?/sec    1.00      6.5±0.07ns        ? ?/sec
date/add_days/sameyear/duration                                 3.18     20.7±0.18ns        ? ?/sec    1.00      6.5±0.05ns        ? ?/sec
date/add_days/sameyear/span                                     4.86     31.7±0.33ns        ? ?/sec    1.00      6.5±0.05ns        ? ?/sec
date/add_years_months_days                                      3.15     31.6±0.35ns        ? ?/sec    1.00     10.0±0.11ns        ? ?/sec
date/day_of_year/get/leap                                       9.94      8.9±0.06ns        ? ?/sec    1.00      0.9±0.01ns        ? ?/sec
date/day_of_year/get/noleap                                     9.95      8.9±0.08ns        ? ?/sec    1.00      0.9±0.01ns        ? ?/sec
date/days_in_month/leap/feb                                     2.70      1.1±0.04ns        ? ?/sec    1.00      0.4±0.01ns        ? ?/sec
date/days_in_month/leap/nofeb                                   2.57      1.0±0.04ns        ? ?/sec    1.00      0.4±0.00ns        ? ?/sec
date/days_in_month/noleap/feb                                   2.70      1.1±0.04ns        ? ?/sec    1.00      0.4±0.03ns        ? ?/sec
date/days_in_month/noleap/nofeb                                 2.50      1.0±0.01ns        ? ?/sec    1.00      0.4±0.00ns        ? ?/sec
date/difference_days/duration                                   3.36      9.4±0.09ns        ? ?/sec    1.00      2.8±0.02ns        ? ?/sec
date/difference_days/span                                       3.68      9.3±0.08ns        ? ?/sec    1.00      2.5±0.02ns        ? ?/sec
date/tomorrow/diff-month                                        15.15    19.0±0.26ns        ? ?/sec    1.00      1.3±0.01ns        ? ?/sec
date/tomorrow/diff-year                                         13.08    18.9±0.17ns        ? ?/sec    1.00      1.4±0.01ns        ? ?/sec
date/tomorrow/same-month                                        24.50    18.9±0.16ns        ? ?/sec    1.00      0.8±0.01ns        ? ?/sec
date/yesterday/diff-month                                       14.86    19.0±0.29ns        ? ?/sec    1.00      1.3±0.01ns        ? ?/sec
date/yesterday/diff-year                                        17.83    18.9±0.10ns        ? ?/sec    1.00      1.1±0.01ns        ? ?/sec
date/yesterday/same-month                                       27.94    18.9±0.19ns        ? ?/sec    1.00      0.7±0.01ns        ? ?/sec
datetime/add_years_months_days                                  2.94     89.4±0.47ns        ? ?/sec    1.00     30.4±0.18ns        ? ?/sec
parse/civil_datetime                                            1.50     35.4±0.28ns        ? ?/sec    1.00     23.5±0.17ns        ? ?/sec
parse/friendly/long/span                                        1.05     92.8±0.30ns        ? ?/sec    1.00     88.0±0.09ns        ? ?/sec
parse/friendly/longer/span                                      1.05     93.7±0.24ns        ? ?/sec    1.00     89.0±0.12ns        ? ?/sec
parse/friendly/longest-time/duration                            1.10     49.3±1.09ns        ? ?/sec    1.00     44.9±0.68ns        ? ?/sec
parse/friendly/longest-time/span                                1.09    112.2±0.93ns        ? ?/sec    1.00    103.3±0.24ns        ? ?/sec
parse/friendly/longest/span                                     1.07    144.7±1.02ns        ? ?/sec    1.00    134.9±0.30ns        ? ?/sec
parse/friendly/medium/span                                      1.05     55.2±0.17ns        ? ?/sec    1.00     52.8±0.06ns        ? ?/sec
parse/friendly/short/duration                                   1.00     16.3±0.62ns        ? ?/sec    1.03     16.8±0.48ns        ? ?/sec
parse/friendly/short/span                                       1.05     44.6±0.14ns        ? ?/sec    1.00     42.6±0.07ns        ? ?/sec
parse/friendly/tiny/duration                                    1.04      8.9±0.25ns        ? ?/sec    1.00      8.6±0.48ns        ? ?/sec
parse/friendly/tiny/span                                        1.04     28.4±0.08ns        ? ?/sec    1.00     27.4±0.10ns        ? ?/sec
parse/rfc2822                                                   1.61     41.9±0.45ns        ? ?/sec    1.00     26.1±0.30ns        ? ?/sec
parse/strptime/oneshot                                          1.25     73.9±0.98ns        ? ?/sec    1.00     59.0±0.94ns        ? ?/sec
print/civil_datetime                                            1.02     51.4±0.26ns        ? ?/sec    1.00     50.5±0.18ns        ? ?/sec
timestamp/add_time_secs/duration                                1.07      2.9±0.03ns        ? ?/sec    1.00      2.7±0.02ns        ? ?/sec
timestamp/add_time_secs/span                                    1.03     11.2±0.10ns        ? ?/sec    1.00     10.9±0.09ns        ? ?/sec
timestamp/add_time_subsec/duration                              1.00      3.1±0.02ns        ? ?/sec    1.02      3.1±0.05ns        ? ?/sec
timestamp/add_time_subsec/span                                  1.00     17.3±0.21ns        ? ?/sec    1.05     18.1±0.21ns        ? ?/sec
timestamp/every_hour_in_week/byhand                             4.10    428.7±3.70ns        ? ?/sec    1.00    104.5±1.31ns        ? ?/sec
timestamp/every_hour_in_week/series                             23.83     2.5±0.01µs        ? ?/sec    1.00    106.7±0.73ns        ? ?/sec
timestamp/to_civil_datetime_offset_conversion                   4.96     21.9±0.17ns        ? ?/sec    1.00      4.4±0.04ns        ? ?/sec
timestamp/to_civil_datetime_offset_holistic                     5.14     21.5±0.17ns        ? ?/sec    1.00      4.2±0.02ns        ? ?/sec
timestamp/to_civil_datetime_static/America-New-York/zoneinfo    3.15     56.0±0.55ns        ? ?/sec    1.00     17.7±0.15ns        ? ?/sec
timestamp/to_civil_datetime_static/Asia-Shanghai/zoneinfo       3.52     25.3±0.20ns        ? ?/sec    1.00      7.2±0.05ns        ? ?/sec
zoned/fixed_offset_add_time/duration                            1.34     28.0±0.27ns        ? ?/sec    1.00     20.9±0.11ns        ? ?/sec
zoned/fixed_offset_add_time/span                                1.42     54.2±0.30ns        ? ?/sec    1.00     38.3±0.15ns        ? ?/sec
zoned/fixed_offset_to_civil_datetime                            1.11      1.0±0.01ns        ? ?/sec    1.00      0.9±0.02ns        ? ?/sec
zoned/fixed_offset_to_timestamp                                 1.00      0.4±0.03ns        ? ?/sec    1.01      0.4±0.01ns        ? ?/sec
```

Many (but _not_ all) are due to switching to
[Neri-Schneider](https://onlinelibrary.wiley.com/doi/full/10.1002/spe.3172), which makes the conversions between Gregorian calendar dates and
Unix epoch days much faster. Significantly so. These conversions are
used everywhere in Jiff. For example, when moving between timestamps
and civil dates. Or in certain `civil::Date` APIs where the calculation
(like getting the weekday of the current date) is much easier when
using the Unix epoch days representation.

Other optimizations are a result of some new tracking being done
inside a `Span`. Specifically, a `Span` now has a bit set internally
that keeps track of which of each of its ten units are non-zero. This
enables fast paths in some cases. For example, when adding a `1.day()`
(which is a `Span`) to a `civil::Date`, Jiff was previously doing a
bunch of work to handle the fully general case of a `Span` containing
non-zero units for all 10 units. Now, in the common case of just adding
days, Jiff does significantly less work.

Similar ideas apply elsewhere. The end result is that Jiff is now
generally faster or roughly competitive with Chrono (I'm only including
benchmarks that make sense for both Chrono and Jiff):

```
$ critcmp update -g '(.*)/(?:jiff|chrono)$'
group                                                           update//chrono                         update//jiff
-----                                                           --------------                         ------------
civil_datetime/add_days/diffyear/duration                       1.00      8.3±0.07ns        ? ?/sec    1.77     14.6±0.29ns        ? ?/sec
civil_datetime/add_days/sameyear/duration                       1.00      6.4±0.07ns        ? ?/sec    2.29     14.6±0.23ns        ? ?/sec
civil_datetime/to_datetime_static/bundled                       1.00     22.3±0.31ns        ? ?/sec
civil_datetime/to_datetime_static/zoneinfo                      1.12     18.2±0.33ns        ? ?/sec    1.00     16.2±0.17ns        ? ?/sec
civil_datetime/to_timestamp_tzdb_lookup/bundled                 1.00     32.0±0.28ns        ? ?/sec
civil_datetime/to_timestamp_tzdb_lookup/zoneinfo                52.41     2.1±0.01µs        ? ?/sec    1.00     40.7±0.17ns        ? ?/sec
date/add_days/diffyear/duration                                 1.00      5.9±0.04ns        ? ?/sec    1.11      6.5±0.04ns        ? ?/sec
date/add_days/sameyear/duration                                 1.00      2.1±0.03ns        ? ?/sec    3.08      6.5±0.05ns        ? ?/sec
date/difference_days/duration                                   1.23      3.4±0.03ns        ? ?/sec    1.00      2.8±0.02ns        ? ?/sec
date/tomorrow/diff-month                                        1.00      0.4±0.00ns        ? ?/sec    3.24      1.3±0.01ns        ? ?/sec
date/tomorrow/diff-year                                         1.24      1.8±0.02ns        ? ?/sec    1.00      1.4±0.01ns        ? ?/sec
date/tomorrow/same-month                                        1.00      0.4±0.02ns        ? ?/sec    1.93      0.8±0.01ns        ? ?/sec
date/yesterday/diff-month                                       1.00      0.4±0.00ns        ? ?/sec    3.31      1.3±0.01ns        ? ?/sec
date/yesterday/diff-year                                        1.96      2.1±0.02ns        ? ?/sec    1.00      1.1±0.01ns        ? ?/sec
date/yesterday/same-month                                       1.00      0.4±0.00ns        ? ?/sec    1.75      0.7±0.01ns        ? ?/sec
parse/civil_datetime                                            3.13     73.6±0.69ns        ? ?/sec    1.00     23.5±0.17ns        ? ?/sec
parse/rfc2822                                                   2.41     62.9±0.35ns        ? ?/sec    1.00     26.1±0.30ns        ? ?/sec
parse/strptime/oneshot                                          2.93    172.8±3.46ns        ? ?/sec    1.00     59.0±0.94ns        ? ?/sec
parse/strptime/prebuilt                                         1.00     91.0±1.25ns        ? ?/sec
print/civil_datetime                                            3.08    155.6±3.81ns        ? ?/sec    1.00     50.5±0.18ns        ? ?/sec
timestamp/add_time_secs/duration                                2.14      5.8±0.04ns        ? ?/sec    1.00      2.7±0.02ns        ? ?/sec
timestamp/add_time_subsec/duration                              1.84      5.8±0.04ns        ? ?/sec    1.00      3.1±0.05ns        ? ?/sec
timestamp/every_hour_in_week/byhand                             15.83  1654.8±2.72ns        ? ?/sec    1.00    104.5±1.31ns        ? ?/sec
timestamp/to_civil_datetime_offset_conversion                   1.61      7.1±0.04ns        ? ?/sec    1.00      4.4±0.04ns        ? ?/sec
timestamp/to_civil_datetime_static/America-New-York/bundled     1.00     21.3±0.12ns        ? ?/sec
timestamp/to_civil_datetime_static/America-New-York/zoneinfo    1.15     20.5±0.18ns        ? ?/sec    1.00     17.7±0.15ns        ? ?/sec
timestamp/to_civil_datetime_static/Asia-Shanghai/bundled        1.00     20.8±0.21ns        ? ?/sec
timestamp/to_civil_datetime_static/Asia-Shanghai/zoneinfo       2.60     18.7±0.09ns        ? ?/sec    1.00      7.2±0.05ns        ? ?/sec
zoned/fixed_offset_add_time/duration                            1.00      6.0±0.03ns        ? ?/sec    3.50     20.9±0.11ns        ? ?/sec
zoned/fixed_offset_to_civil_datetime                            5.92      5.4±0.01ns        ? ?/sec    1.00      0.9±0.02ns        ? ?/sec
zoned/fixed_offset_to_timestamp                                 3.17      1.2±0.01ns        ? ?/sec    1.00      0.4±0.01ns        ? ?/sec
```

And Jiff is now also generally faster than `time`:

```
$ critcmp update -g '(.*)/(?:jiff|time)$'
group                                                           update//time                           update//jiff
-----                                                           ------------                           ------------
civil_datetime/add_days/diffyear/duration                       1.11     16.2±0.14ns        ? ?/sec    1.00     14.6±0.29ns        ? ?/sec
civil_datetime/add_days/sameyear/duration                       1.14     16.5±0.10ns        ? ?/sec    1.00     14.6±0.23ns        ? ?/sec
date/add_days/diffyear/duration                                 1.11      7.3±0.07ns        ? ?/sec    1.00      6.5±0.04ns        ? ?/sec
date/add_days/sameyear/duration                                 1.16      7.6±0.07ns        ? ?/sec    1.00      6.5±0.05ns        ? ?/sec
date/days_in_month/leap/feb                                     12.53     4.9±0.34ns        ? ?/sec    1.00      0.4±0.01ns        ? ?/sec
date/days_in_month/leap/nofeb                                   9.08      3.5±0.34ns        ? ?/sec    1.00      0.4±0.00ns        ? ?/sec
date/days_in_month/noleap/feb                                   11.14     4.4±0.33ns        ? ?/sec    1.00      0.4±0.03ns        ? ?/sec
date/days_in_month/noleap/nofeb                                 8.20      3.2±0.27ns        ? ?/sec    1.00      0.4±0.00ns        ? ?/sec
date/difference_days/duration                                   1.45      4.1±0.04ns        ? ?/sec    1.00      2.8±0.02ns        ? ?/sec
date/tomorrow/diff-month                                        1.00      0.4±0.00ns        ? ?/sec    3.23      1.3±0.01ns        ? ?/sec
date/tomorrow/diff-year                                         1.00      0.5±0.03ns        ? ?/sec    2.90      1.4±0.01ns        ? ?/sec
date/tomorrow/same-month                                        1.00      0.4±0.01ns        ? ?/sec    1.96      0.8±0.01ns        ? ?/sec
date/yesterday/diff-month                                       1.00      0.4±0.00ns        ? ?/sec    3.31      1.3±0.01ns        ? ?/sec
date/yesterday/diff-year                                        1.00      0.7±0.02ns        ? ?/sec    1.55      1.1±0.01ns        ? ?/sec
date/yesterday/same-month                                       1.00      0.4±0.00ns        ? ?/sec    1.76      0.7±0.01ns        ? ?/sec
parse/civil_datetime                                            1.43     33.6±0.19ns        ? ?/sec    1.00     23.5±0.17ns        ? ?/sec
parse/rfc2822                                                   3.12     81.4±1.60ns        ? ?/sec    1.00     26.1±0.30ns        ? ?/sec
parse/strptime/oneshot                                                                                 1.00     59.0±0.94ns        ? ?/sec
parse/strptime/prebuilt                                         1.00    117.0±1.08ns        ? ?/sec
print/civil_datetime                                            1.02     51.5±0.75ns        ? ?/sec    1.00     50.5±0.18ns        ? ?/sec
timestamp/add_time_secs/duration                                7.04     19.0±0.21ns        ? ?/sec    1.00      2.7±0.02ns        ? ?/sec
timestamp/add_time_subsec/duration                              6.06     19.1±0.25ns        ? ?/sec    1.00      3.1±0.05ns        ? ?/sec
timestamp/every_hour_in_week/byhand                             32.29     3.4±0.04µs        ? ?/sec    1.00    104.5±1.31ns        ? ?/sec
timestamp/to_civil_datetime_offset_conversion                   3.30     14.6±0.15ns        ? ?/sec    1.00      4.4±0.04ns        ? ?/sec
timestamp/to_civil_datetime_offset_holistic                     4.48     18.7±0.07ns        ? ?/sec    1.00      4.2±0.02ns        ? ?/sec
zoned/fixed_offset_add_time/duration                            1.11     23.2±0.14ns        ? ?/sec    1.00     20.9±0.11ns        ? ?/sec
zoned/fixed_offset_to_civil_datetime                            1.05      1.0±0.02ns        ? ?/sec    1.00      0.9±0.02ns        ? ?/sec
zoned/fixed_offset_to_timestamp                                 7.01      2.7±0.04ns        ? ?/sec    1.00      0.4±0.01ns        ? ?/sec
```

The benchmarks are lightly commented and meant to be easy
to read. I try not to obscure them with macro bullshit. If
anyone has questions about them, I'm happy to
[answer them in Discussions](https://github.com/BurntSushi/jiff/discussions).

Fixes #235, Fixes #255